refactored schedule grid component in preperation for adding preferences
